breakout getUserNames from startscreen to startscreenUtils
