+++ server/types.ts
@@ -30,3 +30,12 @@ export interface PlayerSocket extends SocketIO.Socket {
 export interface Enemies {
     [key: string]: Enemy
 }
+
+export interface Users {
+    [key: string]: User;
+}
+
+interface User {
+    username: string;
+    role: string;
+}
